requirements/development.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/development.txt requirements/development.in
#
aiohttp==3.4.4
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via aiohttp
attrs==18.2.0             # via aiohttp
backcall==0.1.0           # via ipython
bcrypt==3.1.4             # via paramiko
cffi==1.11.5              # via bcrypt, cryptography, pynacl
chardet==3.0.4            # via aiohttp
click==7.0                # via pip-tools
cryptography==2.4.2       # via paramiko
decorator==4.3.0          # via ipython, traitlets
fabric3==1.14.post1
idna-ssl==1.1.0           # via aiohttp
idna==2.7                 # via cryptography, idna-ssl, yarl
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.1.1
jedi==0.13.1              # via ipython
multidict==4.5.2          # via aiohttp, yarl
paramiko==2.4.2           # via fabric3
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.1.0
prompt-toolkit==2.0.7     # via ipython
ptyprocess==0.6.0         # via pexpect
pyasn1==0.4.4             # via paramiko
pycparser==2.19           # via cffi
pygments==2.3.0           # via ipython
pynacl==1.3.0             # via paramiko
python-box==3.2.3
six==1.11.0               # via bcrypt, cryptography, fabric3, pip-tools, prompt-toolkit, pynacl, traitlets
traitlets==4.3.2          # via ipython
wcwidth==0.1.7            # via prompt-toolkit
yarl==1.2.6               # via aiohttp
